Embed ansible setup collector
Instead of using setup or gather facts, the module calls
the collectors directly to avoid passing the whole of
ansible_facts into the module. This approach makes the
size of the resulting job output considerably smaller!

In addition a check has been added to ignore disks that
are on a USB bus.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>